Trying to fix "Docker Attach" no longer working to execute commands.
No idea if this is the proper fix. But when I readded -it to the Extra Parameters in the Unraid Template it allowed me to send commands to the server via the Docker Attach command. I think this would be the way to fix it long term for new people adding the app via the Unraid App Store but I honestly have no idea so please verify first.